backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.36
# SQLAlchemy dialect for libSQL/Turso
sqlalchemy-libsql==0.1.0
# psycopg2-binary==2.9.9  # PostgreSQL - replaced with SQLite
alembic==1.12.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic==2.9.2
pandas==2.2.2
python-dotenv==1.0.0
email-validator==2.1.1